Updating repo generation stage
Previous implementation had a fatal flaw: `straight-recipes-retrieve` is
only called when `straight-use-package` is called with a plain package
name and there is no package override is registered.

When a package is overridden or a plist is given to
`straight-use-package` then the recipe is assumed to be complete and
`straight-recipes-retrieve` call is skipped.

New approach is to register our own override via
`straight-override-recipe` for all packages so the recipe retrieval step
is skipped so is the whole cloning stage because `:local-repo` is
already set now.

Logging style is also updated to be distinguishable from straight's own logs.